add attendance crud structure
